Improve latency on Ubuntu (#1)
When sending an array of bytes, the WrapperByteChannel will first send a 4-byte message containing the length of the array, and then send the array. On Linux, the receiving thread reads the 4-byte message first, and then tries to receive the next message. But the next message hasn't arrived yet, so the thread gets descheduled---causing a 50 millisecond penalty!

No idea why the context switch cost is so high, but I can empirically confirm that the first message is received within a millisecond or so and the second message takes around 50 milliseconds. The problem doesn't happen if you send just one message instead of two!

I fixed the behavior by combining the 4-byte prefix with the serialized message into a single byte buffer.